Release IS48

+++ source/backend/api/Services/AcquisitionFileService.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Security.Claims;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using Pims.Api.Helpers.Exceptions;
 using Pims.Core.Exceptions;
 using Pims.Core.Extensions;
 using Pims.Dal.Constants;
@@ -99,8 +100,14 @@ namespace Pims.Api.Services
             _logger.LogInformation("Adding acquisition file with id {id}", acquisitionFile.Internal_Id);
             _user.ThrowIfNotAuthorized(Permissions.AcquisitionFileAdd);
 
-            acquisitionFile.AcquisitionFileStatusTypeCode = "ACTIVE";
+            // validate the new acq region
+            var cannotDetermineRegion = _lookupRepository.GetAllRegions().FirstOrDefault(x => x.RegionName == "Cannot determine");
+            if (acquisitionFile.RegionCode == cannotDetermineRegion.RegionCode)
+            {
+                throw new BadRequestException("Cannot set an acquisition file's region to 'cannot determine'");
+            }
 
+            acquisitionFile.AcquisitionFileStatusTypeCode = "ACTIVE";
             MatchProperties(acquisitionFile);
 
             var newAcqFile = _acqFileRepository.Add(acquisitionFile);
@@ -120,6 +127,13 @@ namespace Pims.Api.Services
             if (!userOverride)
             {
                 ValidateMinistryRegion(acquisitionFile.Internal_Id, acquisitionFile.RegionCode);
+            }
+
+            // reset the region
+            var cannotDetermineRegion = _lookupRepository.GetAllRegions().FirstOrDefault(x => x.RegionName == "Cannot determine");
+            if (acquisitionFile.RegionCode == cannotDetermineRegion.RegionCode)
+            {
+                throw new BadRequestException("Cannot set an acquisition file's region to 'cannot determine'");
             }
 
             var newAcqFile = _acqFileRepository.Update(acquisitionFile);
@@ -164,6 +178,11 @@ namespace Pims.Api.Services
             List<PimsPropertyAcquisitionFile> differenceSet = currentProperties.Where(x => !acquisitionFile.PimsPropertyAcquisitionFiles.Any(y => y.Internal_Id == x.Internal_Id)).ToList();
             foreach (var deletedProperty in differenceSet)
             {
+                var acqFileProperties = _acquisitionFilePropertyRepository.GetPropertiesByAcquisitionFileId(acquisitionFile.Internal_Id).FirstOrDefault(ap => ap.PropertyId == deletedProperty.PropertyId);
+                if (acqFileProperties.PimsActInstPropAcqFiles.Any() || acqFileProperties.PimsTakes.Any())
+                {
+                    throw new BusinessRuleViolationException();
+                }
                 _acquisitionFilePropertyRepository.Delete(deletedProperty);
                 if (deletedProperty.Property.IsPropertyOfInterest.HasValue && deletedProperty.Property.IsPropertyOfInterest.Value)
                 {
